[IMP] finished computing serial number automatically using VIN
A module was created to give better functionality to serial numbers. Now, the serial number is automatically computed by using the VIN from the motorcycle product model. By inheriting from 'stock.lot', we get access to both the 'name' field to change the Lot/Serial Number, as well as the necessary data to get the VIN. An XML file with the serial numbers sequence parameters was created.