[IMP] GE09-TEAM6 : Add  motorcycle specific info to the repair view and create a smart button that makes a relation between the montorcyclke registry to the repair order
We added a vin field to the repair view to compute specific values related to it, after that we make a relationship between motorcycle registry and repair order, compute a variable
that counts how many repair orders there are then we created a function to show the tree view of the repair orders with a domain to only show the repair orders that we expect

[IMP] GE09-TEAM6: 'lot_id' is now automatically obtained from motorcycle registry.

This change facilitates the automation of processes like creating a repair through a portal.

[IMP] GE09-TEAM6: Resolved formatting issues

Changed double quotes to single quotes to maintain consistency between files. Also, removed trailing whitespace from line breaks at the end of some files